Remove pre-comp section, update f3dex2 description
I've spoken with Wiseguy about this, and his precomp process is kind of technical and specific and doesn't really belong on a general list for N64 resources, at least without sufficient explanation and context. This PR also removes another game-specific link in there which is not one of Wiseguy's precomps, so I'm not sure if that needs to go elsewhere or what.